+++ Dyna/Cards/BunkerCard.cs
@@ -18,8 +18,7 @@ internal sealed class BunkerCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = StableSpr.cards_BigShield,
-			//Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Bunker.png")).Sprite,
+			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Bunker.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Bunker", "name"]).Localize
 		});
 	}