+++ Dyna/Cards/IncomingCard.cs
@@ -18,8 +18,7 @@ internal sealed class IncomingCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = StableSpr.cards_Cannon,
-			//Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Incoming.png")).Sprite,
+			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Incoming.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Incoming", "name"]).Localize
 		});
 	}