+++ Dyna/Cards/MegatonBlastCard.cs
@@ -18,8 +18,7 @@ internal sealed class MegatonBlastCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = StableSpr.cards_AggressiveArmoring,
-			//Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/MegatonBlast.png")).Sprite,
+			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/MegatonBlast.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "MegatonBlast", "name"]).Localize
 		});
 	}