+++ Dyna/Cards/NitroCard.cs
@@ -18,8 +18,7 @@ internal sealed class NitroCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = StableSpr.cards_Overpower,
-			//Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Nitro.png")).Sprite,
+			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/Nitro.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Nitro", "name"]).Localize
 		});
 	}