+++ Dyna/Cards/RemoteDetonatorCard.cs
@@ -19,8 +19,7 @@ internal sealed class RemoteDetonatorCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = StableSpr.cards_hacker,
-			//Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/RemoteDetonator.png")).Sprite,
+			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/RemoteDetonator.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "RemoteDetonator", "name"]).Localize
 		});
 	}