+++ Dyna/Cards/SwiftChargeCard.cs
@@ -18,8 +18,7 @@ internal sealed class SwiftChargeCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = StableSpr.cards_GoatDrone,
-			//Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/SwiftCharge.png")).Sprite,
+			Art = helper.Content.Sprites.RegisterSprite(package.PackageRoot.GetRelativeFile("assets/Cards/SwiftCharge.png")).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "SwiftCharge", "name"]).Localize
 		});
 	}